Make all the doubles into floats.

+++ src/main/java/com/github/tommyettinger/artsi/Node2DImpl.java
@@ -9,10 +9,10 @@ import java.util.List;
  */
 class Node2DImpl extends Node2D {
 
-    double minX = Double.MAX_VALUE;
-    double minY = Double.MAX_VALUE;
-    double maxX = -Double.MAX_VALUE;
-    double maxY = -Double.MAX_VALUE;
+    float minX = Float.MAX_VALUE;
+    float minY = Float.MAX_VALUE;
+    float maxX = -Float.MAX_VALUE;
+    float maxY = -Float.MAX_VALUE;
 
     /**
      * Create a rectangular data node
@@ -22,7 +22,7 @@ class Node2DImpl extends Node2D {
      * @param maxX the minimum y component of the data
      * @param maxY the maximum y component of the data
      */
-    protected Node2DImpl(double minX, double minY, double maxX, double maxY) {
+    protected Node2DImpl(float minX, float minY, float maxX, float maxY) {
         super(true);
         set(minX, minY, maxX, maxY);
     }
@@ -36,7 +36,7 @@ class Node2DImpl extends Node2D {
     }
 
     @Override
-    protected final void set(double minX, double minY, double maxX, double maxY) {
+    protected final void set(float minX, float minY, float maxX, float maxY) {
         this.minX = minX;
         this.minY = minY;
         this.maxX = maxX;
@@ -44,22 +44,22 @@ class Node2DImpl extends Node2D {
     }
 
     @Override
-    public double getMinX() {
+    public float getMinX() {
         return minX;
     }
 
     @Override
-    public double getMinY() {
+    public float getMinY() {
         return minY;
     }
 
     @Override
-    public double getMaxX() {
+    public float getMaxX() {
         return maxX;
     }
 
     @Override
-    public double getMaxY() {
+    public float getMaxY() {
         return maxY;
     }
 
@@ -71,7 +71,7 @@ class Node2DImpl extends Node2D {
      * @return an int representing whether a is greater/lesser/equal to the b
      */
     static int compareMinX(Node2D a, Node2D b) {
-        return Double.compare(a.getMinX(), b.getMinX());
+        return Float.compare(a.getMinX(), b.getMinX());
     }
 
     /**
@@ -82,7 +82,7 @@ class Node2DImpl extends Node2D {
      * @return an int representing whether a is greater/lesser/equal to the b
      */
     static int compareMinY(Node2D a, Node2D b) {
-        return Double.compare(a.getMinY(), b.getMinY());
+        return Float.compare(a.getMinY(), b.getMinY());
     }
 
     /**
@@ -92,7 +92,7 @@ class Node2DImpl extends Node2D {
      * @param b node b
      * @return the area of performing a union with nodes a and b
      */
-    static double enlargedArea(Node2D a, Node2D b) {
+    static float enlargedArea(Node2D a, Node2D b) {
         return enlargedArea(a.getMinX(), a.getMinY(), a.getMaxX(), a.getMaxY(), b.getMinX(), b.getMinY(), b.getMaxX(), b.getMaxY());
     }
 
@@ -103,7 +103,7 @@ class Node2DImpl extends Node2D {
      * @param b node b
      * @return the area of intersecting a with b
      */
-    static double intersectionArea(Node2D a, Node2D b) {
+    static float intersectionArea(Node2D a, Node2D b) {
         return intersectionArea(a.getMinX(), a.getMinY(), a.getMaxX(), a.getMaxY(), b.getMinX(), b.getMinY(), b.getMaxX(), b.getMaxY());
     }
 
@@ -112,64 +112,64 @@ class Node2DImpl extends Node2D {
     }
 
     static boolean contains(Node2D a,
-                            double bMinX,
-                            double bMinY,
-                            double bMaxX,
-                            double bMaxY) {
+                            float bMinX,
+                            float bMinY,
+                            float bMaxX,
+                            float bMaxY) {
         return contains(a.getMinX(), a.getMinY(), a.getMaxX(), a.getMaxY(), bMinX, bMinY, bMaxX, bMaxY);
     }
 
-    static double enlargedArea(
-            double aMinX,
-            double aMinY,
-            double aMaxX,
-            double aMaxY,
-            double bMinX,
-            double bMinY,
-            double bMaxX,
-            double bMaxY
+    static float enlargedArea(
+            float aMinX,
+            float aMinY,
+            float aMaxX,
+            float aMaxY,
+            float bMinX,
+            float bMinY,
+            float bMaxX,
+            float bMaxY
     ) {
         return (Math.max(bMaxX, aMaxX) - Math.min(bMinX, aMinX)) *
                 (Math.max(bMaxY, aMaxY) - Math.min(bMinY, aMinY));
     }
 
-    static double intersectionArea(
-            double aMinX,
-            double aMinY,
-            double aMaxX,
-            double aMaxY,
-            double bMinX,
-            double bMinY,
-            double bMaxX,
-            double bMaxY
+    static float intersectionArea(
+            float aMinX,
+            float aMinY,
+            float aMaxX,
+            float aMaxY,
+            float bMinX,
+            float bMinY,
+            float bMaxX,
+            float bMaxY
     ) {
-        double minX = Math.max(aMinX, bMinX);
-        double minY = Math.max(aMinY, bMinY);
-        double maxX = Math.min(aMaxX, bMaxX);
-        double maxY = Math.min(aMaxY, bMaxY);
+        float minX = Math.max(aMinX, bMinX);
+        float minY = Math.max(aMinY, bMinY);
+        float maxX = Math.min(aMaxX, bMaxX);
+        float maxY = Math.min(aMaxY, bMaxY);
 
         return Math.max(0, maxX - minX) *
                 Math.max(0, maxY - minY);
     }
 
     static boolean contains(
-            double aMinX,
-            double aMinY,
-            double aMaxX,
-            double aMaxY,
+            float aMinX,
+            float aMinY,
+            float aMaxX,
+            float aMaxY,
             Node2D b) {
         return contains(aMinX, aMinY, aMaxX, aMaxY, b.getMinX(), b.getMinY(), b.getMaxX(), b.getMaxY());
     }
 
     static boolean contains(
-            double aMinX,
-            double aMinY,
-            double aMaxX,
-            double aMaxY,
-            double bMinX,
-            double bMinY,
-            double bMaxX,
-            double bMaxY
+            float aMinX,
+            float aMinY,
+            float aMaxX,
+            float aMaxY,
+            float bMinX,
+            float bMinY,
+            float bMaxX,
+            float bMaxY
     ) {
         return aMinX <= bMinX &&
                 aMinY <= bMinY &&
@@ -196,10 +196,10 @@ class Node2DImpl extends Node2D {
      * @param b     the other bound
      * @return whether the two bounds intersect
      */
-    public static boolean intersects(double aMinX,
-                                     double aMinY,
-                                     double aMaxX,
-                                     double aMaxY
+    public static boolean intersects(float aMinX,
+                                     float aMinY,
+                                     float aMaxX,
+                                     float aMaxY
             , Node2D b) {
         return intersects(aMinX, aMinY, aMaxX, aMaxY, b.getMinX(), b.getMinY(), b.getMaxX(), b.getMaxY());
     }
@@ -216,14 +216,14 @@ class Node2DImpl extends Node2D {
      * @return whether the two bounds intersect
      */
     public static boolean intersects(
-            double aMinX,
-            double aMinY,
-            double aMaxX,
-            double aMaxY,
-            double bMinX,
-            double bMinY,
-            double bMaxX,
-            double bMaxY
+            float aMinX,
+            float aMinY,
+            float aMaxX,
+            float aMaxY,
+            float bMinX,
+            float bMinY,
+            float bMaxX,
+            float bMaxY
 
     ) {
         return bMinX <= aMaxX &&
